requirements.txt
# Backend & API
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# ML & Data Processing
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
joblib==1.3.2

# Experiment Tracking
wandb==0.16.1

# Monitoring
prometheus-client==0.19.0

# Frontend
streamlit==1.29.0
requests==2.31.0

# Testing & Code Quality
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.1
flake8==6.1.0
pylint==3.0.3
mypy==1.7.1

# Development
python-dotenv==1.0.0
